refactor: merge TreeLandHelper to Helper
doesn't need split two helper class.

Log:
