+++ src/treeland/quick/protocols/extforeigntoplevellist.h
@@ -5,7 +5,7 @@
 
 #include "ext-foreign-toplevel-list-server-protocol.h"
 #include "ext_foreign_toplevel_list_impl.h"
-#include "treelandhelper.h"
+#include "helper.h"
 
 #include <wxdgsurface.h>
 #include <wquickwaylandserver.h>
